requirements.txt
asgiref==3.8.1
Django==5.2.2
sqlparse==0.5.3
django-allauth==65.9.0
django-widget-tweaks==1.5.0
django-debug-toolbar==5.2.0
pytest==8.4.0
pytest-cov==6.2.1
django-rest-framework==0.1.0
pytest-django==4.11.1